Bayesian coin flip with Gaussian prior
Coin flip simulator of a biased coin. Uses Bayesian inference to iteratively compute the posterior distribution of the "Heads" as more coin flips are added to the pool of data. Includes an animation of the evolution of the posterior distribution and a trace plot of the location of its maximum at each number of coin flips