1D Gaussian random surface and electromagnetics
A function for simulating a spatially smooth 1D random Gaussian surface (e.g. rough surface) and computing the effective permittivities of 2 equivalent homogeneous layers (above and below zero) using the Maxwell-Garnett model of effective medium theory. This can be used for computations related to electromagnetic modelling, for example approximating the reflectance/transmittance properties of rough surfaces, or for use in multilayer structures, etc.